Tested player controllers
I have created a few iterations of the player controller. I had a bunch of problems with ground detection with the inbuilt character-controller, so I made a rigidbody based controller. That controller had even more problems so I switched back to a inbuilt character-controller based controller, with a new and improved 'ground magnet' that forces the player to actually attach itself to the ground (fixing a lot of the problems). Soo that is probably the approach I will use for the final game.