Added moving platforms
right now the speed of a platform is not constant (even with linear easing) when using bezier curves. The closer to a constant gradient, the more accurate it will be. need to make moving platforms factor in bezier curve derivative to fix this.
Also player's 'sticking' to platforms is dependant on framerate at the moment- it will work worse the lower the framerate. This is because of a delay between moving the platform and moving the player.
Also player's velocity does not update to match moving platform speed after leaving platform
All these things should be fixed in the future.