Updated player and moving platforms
moving platforms now have a constant velocity (on 'None' ease type). Player now attaches to moving platforms (the camera does not like it, unfortunately, due to the fact the player now moves on fixedupdate AND update. this will be fixed later maybe).

There is also now an 'automatic' bezier curve thing, although it might be a bit broken