introduced a bunch of bugs and Player now updates on fixedupdate
this is a big change that needs to happen to stop unpredictable behaviour and make physics interactions stable. unfortunately it also introduces (and reveals) a bunch of bugs.
-jitter
-slope bounce sometimes (especially while rolling)
-going up slopes that 'scrape' against the collider slows down the player
-rolling has some weird behaviour  
-ground magnet is now too good at its job sometimes
-player on moving platform has a noticeable movement delay
-player jitters on looping moving platform when it restarts a loop

hopefully these can all be fixed in time.
Oh, i also added player camera looking, which was really easy, but only if you set the player animations to humanoid, which broke some of the animations slightly. sooo that's not good.